requirements.txt
aiogram
requests
python-dotenv
APScheduler